Create a more thread-safe implementation of Task.
The UI thread can mutate weights and parameters of a Task object arbitrarily without locking.
This means that values can change while rollouts are happening, or while the residual function is evaluated.

The new threading model is:
1. Each task should define a ResidualFn object which is used for evaluating the residuals.
2. At the start of the planning iteration, the Agent takes a copy of this ResidualFn object, and it's used for all rollout threads. Rollout threads don't need to lock access to that instance.
3. There's one instance of the ResidualFn per task that is used by the physics thread and for plotting, and that one is updated from the latest UI values on every call to Transition or Reset, with a lock.

There is still thread-unsafety in that the UI can mutate the weights while they are copied into the ResidualFn.

For backwards compatibility, existing task implementations continue to operate as they did before this change, by using a ResidualFn implementation that simply uses the weights and parameters from the Task object directly.

PiperOrigin-RevId: 530621190
Change-Id: Id7859b708717b5676ef1810a37c89d15129f8d72