Add an option to the CEM planner to keep exploring after the variance has converged to a small amount.
At the moment, as CEM finds a good solution, the variance of the samples it produces goes down to zero. This means that when a big external change happens, there is not enough exploration noise to find a new solution.
After this commit, there's a new option, explore_fraction, which keeps a fraction of the rollout trajectories using the initial exploration noise, instead of the one derived from CEM.

Switch the Shadow task to use CEM.

PiperOrigin-RevId: 696509896
Change-Id: Ib8ad7cda058da1dc7a82c7dff3fcf33a8dcdab2f